Moved SimulationHelper and generator files to dedicated simulation folder
